practice/testing_allur_server/requirements.txt
pycef
click
ldap3
slapd
grpcio
pytest
urllib3
aiohttp
pexpect
tenacity
requests
xmltodict
dicttoxml
mergedeep
sqlalchemy
websockets
ruamel.yaml
pytest-faker
grpcio-tools
pytest-check
pytest-xdist
pytest-repeat
pytest-randomly
google-crc32c
allure-pytest
testresources
pyshark==0.4.6
pydantic==1.10.13
python-telegram
psycopg2-binary
pytest-testrail
protobuf==3.20.0
paho-mqtt==1.6.1
python-ldap-test
teamcity-messages
pytest-rerunfailures
robotframework==4.0.3
allure-python-commons


